+++ NEWS.md
@@ -4,6 +4,8 @@
   The default `msg.sender` (0x000...000) passed the check unless an address was authorised on the gold contract, in which case the transaction reverted. 
   Fixed by setting the `from` field to `summoner.owner` when checking gold claims.
 
+- Summoners are now only sent to the cellar if their expected loot is >= 5.
+
 # raritybot 1.3.2
 
 - Start implementing crafting: new command `show crafting-proba -n DC` will show crafting probabilities for an item of a given DC (20 by default) for a range of INT and craft levels.